[hotfix][core] Merge ClassLoaderUtils classes
