--- flink-core/src/main/java/org/apache/flink/util/ClassLoaderUtils.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.util;
-
-import java.io.File;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Arrays;
-
-/**
- * Collection of class-loading utilities.
- */
-public final class ClassLoaderUtils {
-
-	public static URL[] getClasspathURLs() {
-		final String[] cp = System.getProperty("java.class.path").split(File.pathSeparator);
-
-		return Arrays.stream(cp)
-			.filter(str -> !str.isEmpty())
-			.map(ClassLoaderUtils::parse)
-			.toArray(URL[]::new);
-	}
-
-	private static URL parse(String fileName) {
-		try {
-			return new File(fileName).toURI().toURL();
-		} catch (MalformedURLException e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	private ClassLoaderUtils() {}
-}
